✅ Test AWS2 SQS in isolation Fixes #2606
