+++ integration-test-groups/aws2-quarkus-client/aws2-sqs-sns/src/test/resources/META-INF/services/org.apache.camel.quarkus.test.support.aws2.Aws2TestEnvCustomizer
@@ -1,1 +1,1 @@
-org.apache.camel.quarkus.component.aws2.sqs.it.Aws2SqsQuarkusClientTestEnvCustomizer
+org.apache.camel.quarkus.component.aws2.sns.it.Aws2SnsQuarkusClientTestEnvCustomizer
