+++ integration-test-groups/aws2/aws2-sqs-sns/pom.xml
@@ -64,11 +64,6 @@
             <artifactId>camel-quarkus-integration-tests-support-aws2</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.awaitility</groupId>
-            <artifactId>awaitility</artifactId>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <profiles>
