+++ integration-test-groups/aws2/aws2-sqs/src/test/resources/META-INF/services/org.apache.camel.quarkus.test.support.aws2.Aws2TestEnvCustomizer
@@ -1,0 +1,1 @@
+org.apache.camel.quarkus.component.aws2.sqs.it.Aws2SqsTestEnvCustomizer
