+++ integration-test-groups/aws2/pom.xml
@@ -43,6 +43,7 @@
         <module>aws2-lambda</module>
         <module>aws2-s3</module>
         <module>aws2-ses</module>
+        <module>aws2-sqs</module>
         <module>aws2-sqs-sns</module>
     </modules>
